refactor(week-3): moved testing utils to testing_utils file
